add tmap
