--- app/src/main/java/com/gitturami/bike/MapActivity.kt
@@ -1,48 +1,0 @@
-package com.gitturami.bike
-
-import android.graphics.Color
-import android.os.Bundle
-import android.widget.LinearLayout
-import androidx.appcompat.app.AppCompatActivity
-import com.skt.Tmap.TMapData
-import com.skt.Tmap.TMapPoint
-import com.skt.Tmap.TMapPolyLine
-import com.skt.Tmap.TMapView
-
-class MapActivity : AppCompatActivity() {
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_map)
-
-        val linearLayoutTmap = findViewById(R.id.linearLayoutTmap) as LinearLayout
-        val tMapView = TMapView(this)
-
-        tMapView.setSKTMapApiKey("3826adf1-c99a-4e69-95c5-3763539465ea")
-        linearLayoutTmap.addView(tMapView)
-
-        tMapView.setCenterPoint(126.9573662, 37.5048935)
-
-        val tMapPointStart = TMapPoint(37.5048935, 126.9573662) // 중앙대학교
-        val tMapPointEnd = TMapPoint(37.5099724, 126.9949061) // 반포한강공원
-
-        try {
-            val data = TMapData()
-            data.findPathData(tMapPointStart, tMapPointEnd, object: TMapData.FindPathDataListenerCallback {
-                override fun onFindPathData(path: TMapPolyLine){
-                    runOnUiThread(object: Runnable {
-                        override fun run(){
-                            path.setLineWidth(5f)
-                            path.setLineColor(Color.BLUE)
-                            tMapView.addTMapPath(path)
-                        }
-                    })
-                }
-            })
-
-        } catch (e: Exception) {
-            e.printStackTrace()
-        }
-
-    }
-}
