+++ app/src/main/java/com/gitturami/bike/view/main/presenter/MainContact.kt
@@ -1,10 +1,12 @@
 package com.gitturami.bike.view.main.presenter
 
+import android.widget.LinearLayout
+import com.skt.Tmap.TMapView
+
 interface MainContact {
     interface View {
-
     }
     interface Presenter {
-        fun takeView(view: View)
+        fun takeView(view: LinearLayout)
     }
 }
